Convert QLowEnergyHandle from quint32 to quint16
This is in accordence with the maximal possible handle size
of 0xffff as per Bluetooth spec.

As a side effect processReply() had to be modified since
endHandle (0xffff) + 1 caused a type overrun.

+++ src/bluetooth/qbluetooth.h
@@ -59,7 +59,7 @@ Q_DECLARE_FLAGS(SecurityFlags, Security)
 Q_DECLARE_OPERATORS_FOR_FLAGS(SecurityFlags)
 }
 
-typedef quint32 QLowEnergyHandle;
+typedef quint16 QLowEnergyHandle;
 
 QT_END_NAMESPACE
